requirements.txt
django >3.2,<3.3
djangorestframework
djangorestframework-simplejwt
pytest
pytest-django
celery
redis
django-redis
ffmpeg-python